Es 17/implement pending unit testing (#23)
* Making test api public

* Mocking security configurations for controllers unit testing

+++ src/test/java/com/sabeer/electronic/store/services/UserServiceTest.java
@@ -131,13 +131,17 @@ public class UserServiceTest {
         FileSystem fileSys = FileSystems.getDefault();
         Path originalFilePath = fileSys.getPath(imagePath +"/user_abc.png");
         Path tempFilePath = fileSys.getPath(imagePath + "/user_temp.png");
-        Files.copy(originalFilePath, tempFilePath);
+        if (!tempFilePath.toFile().exists()) {
+            Files.copy(originalFilePath, tempFilePath);
+        }
 
         userService.deleteUser(userId);
 
         Mockito.verify(userRepository, Mockito.times(1)).delete(user);
 
-        Files.copy(tempFilePath, originalFilePath);
+        if (!originalFilePath.toFile().exists()) {
+            Files.copy(tempFilePath, originalFilePath);
+        }
         Files.delete(tempFilePath);
     }
